backend/requirements.txt
# FastAPI 및 웹 서버 관련 패키지
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 이미지 처리 및 파일 업로드 관련
python-multipart==0.0.6
Pillow==10.1.0

# AI/ML 관련 패키지 (PyTorch)
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.3

# HTTP 요청 및 CORS 처리
httpx==0.25.2

# 기타 유틸리티
python-dotenv==1.0.0
